ultimos cambios

+++ include/game_struct.h
@@ -1,12 +1,9 @@
 #ifndef GAME_STRUCT_H_
 #define GAME_STRUCT_H_
-
 #include <stdbool.h>
-
 enum ESTADOS{DERROTA, VICTORIA, JUGAR, SALIR, REINICIAR, FIN};
 enum ESTADOS_BALA{PLAY, RESET, WAIT};
 enum BLOCK_TYPE{NORMAL, HEAVY, UNBREAKABLE};
-
 typedef struct{
 	int state;
 	int points;
@@ -16,32 +13,28 @@ typedef struct{
     int block_counter;
     double timer, prev_timer;
 } GAME_STATE;
-
 typedef struct{
-	int x;
-	int y;
+	float x;   // cambiado de int a float para acumulación precisa con delta time
+	float y;   // cambiado de int a float para acumulación precisa con delta time
     int width;
     int height;
 } PLAYER;
-
 typedef struct{
 	float x, y;
-    int size;
-    float prev_x, prev_y; // Para almacenar la posición anterior
+    float size;
+    float prev_x, prev_y;
     int state;
     float vx, vy;
     float speed_factor;
     float speed;
 } BALL;
-
 typedef struct{
-	int x, y;
+	float x, y;
 	int width, height;
 	int dureza;
 	bool alive;
     int type;
 } BLOCK;
-
 typedef enum {
     COLOR_RED = 0,
     COLOR_ORANGE,
@@ -54,8 +47,4 @@ typedef enum {
     COLOR_WHITE,
     COLOR_GRAY
 } BlockColor;
-
-
-
-
 #endif /* GAME_STRUCT_H_ */
